Extract garden to BalfurStage component
