+++ src/components/BalfurStage/index.js
@@ -1,0 +1,3 @@
+import BalfurStage from './BalfurStage';
+
+export default BalfurStage;
